`SKTestSession`: finish all unfinished transactions when beginning tests
Filed this as `FB11799675`. I noticed some tests began with prior transactions (despite using `SKTestSession.clearTransactions`), so this workaround ensures each test begins in a clean state.
